Match any branch name even if it doesn't follow convention
This is needed to workaround [nebula-plugins/nebula-release-plugin#93](https://github.com/nebula-plugins/nebula-release-plugin/issues/93).
The plugin currently fails if a branch doesn't match the pattern.

With this change any branch name will match, if it follows convention
the match will return only the issue ID, if it doesn't it will return
the full branch name, which may potentially fail if submitting to the
portal, but won't break other builds.